+++ src/app/src/components/messageTypes/bbcode/__all__.tsx
@@ -1,0 +1,8 @@
+import { registerBBCodeTag } from '@src/shared/components/bbcode/parser';
+import ImageTag from './ImageTag';
+import UrlTag from './UrlTag';
+import PlainText from './PlainText';
+
+registerBBCodeTag('_text', PlainText);
+registerBBCodeTag('img', ImageTag);
+registerBBCodeTag('url', UrlTag);
